refactor: stores 폴더 생성
