+++ src/App.jsx
@@ -2,7 +2,7 @@ import { ThemeProvider } from "styled-components";
 import { theme } from "./style";
 
 import { Provider } from "react-redux";
-import { productInfoListStore } from "./productInfoListStore";
+import { productInfoListStore } from "./stores/productInfoListStore";
 
 import Header from "./components/Header";
 import Main from "./components/Main";
